Update P25Hosts.txt
Line# 147 updated to remove the word county to shorten the description.  Line# 642/643 were deleted as this talkgroup 31267 is depreciated and no longer operational.  This is a resubmission of these due to an error in removing the talkgroup details from the description; however, looking through the file, there are many like this and it's really an unnecessary duplication of this information when it's viewed on Pi-Star's website and other resources.